requirements.txt
streamlit>=1.28
spacy>=3.7
pdfplumber>=0.9.0
python-docx>=0.8.11
fpdf2>=2.7.0
reportlab>=4.0.0
nltk>=3.8
pandas>=2.0
numpy>=1.24
google-generativeai>=0.3.0
python-dotenv>=1.0.0
